Changed selectLast(n: Long) to selectLast(n: Int) because of internal limitations of list size

+++ common/databaseApi/src/commonMain/kotlin/dev/oltersdorf/racketclash/database/api/TableBase.kt
@@ -9,7 +9,7 @@ interface TableBase<Item, Filter, Sorting> {
 
     suspend fun selectSingle(id: Long): Item?
 
-    suspend fun selectLast(n: Long): List<Item>
+    suspend fun selectLast(n: Int): List<Item>
 
     suspend fun selectSortedAndFiltered(
         filter: Filter,
